kdecore/date: Do not ignore UTC offset when parsing date strings.
Jon Severinsson noted an error parsing the time zone entry on some
dates, causing the parser to ignore the +/- sign, and contributed a fix
and corresponding test cases.

The patch was reviewed by myself and David Jarvie. I did not find any
open bugs which appeared to be fixed with this patch.